+++ src/module/animations.ts
@@ -1,6 +1,6 @@
 //@ts-ignore
 
-import { TokenData } from '@league-of-foundry-developers/foundry-vtt-types/src/foundry/common/data/module.mjs';
+import type { TokenData } from '@league-of-foundry-developers/foundry-vtt-types/src/foundry/common/data/module.mjs';
 import CONSTANTS from './constants';
 
 export const ANIMATIONS = {